+++ src/components/account/AccountCard/index.tsx
@@ -4,7 +4,6 @@ import {
   Typography,
   Card,
   CardContent,
-  Button,
   Stack,
   Skeleton,
 } from '@mui/material';
