+++ src/services/repositories/transformer.ts
@@ -1,5 +1,6 @@
 import { Asset } from '@/domain/entities/Asset';
-import { AssetNode } from './AssetGraphRepo';
+import { AssetNode, IdentityNode } from './types';
+import { Identity } from '@/domain/entities/Identity';
 
 export function transformAssetNodeToAsset(assetNode: AssetNode): Asset {
   return {
@@ -11,5 +12,26 @@ export function transformAssetNodeToAsset(assetNode: AssetNode): Asset {
     holders: assetNode.holders.totalCount.toString(),
     createdAt: new Date(assetNode.createdAt),
     documents: assetNode.documents.totalCount.toString(),
+  };
+}
+
+export function transformToIdentity(node: IdentityNode): Identity {
+  return {
+    did: node.did,
+    primaryAccount: node.primaryAccount,
+    secondaryAccounts: node.secondaryAccounts.nodes
+      .filter((account) => account.address !== node.primaryAccount)
+      .map((account) => account.address),
+    createdAt: new Date(node.createdAt),
+    claimsCount: node.claimsByTargetId.totalCount,
+    assetsCount: node.heldAssets.totalCount,
+    venuesCount: node.venuesByOwnerId.totalCount,
+    portfoliosCount: node.portfolios.totalCount,
+    ownedAssets: node.assetsByOwnerId.nodes.map((asset) =>
+      transformAssetNodeToAsset(asset),
+    ),
+    heldAssets: node.heldAssets.nodes.map((heldAsset) =>
+      transformAssetNodeToAsset(heldAsset.asset),
+    ),
   };
 }
